+++ beanie/__init__.py
@@ -28,7 +28,7 @@ from beanie.odm.documents import Document
 from beanie.odm.views import View
 from beanie.odm.union_doc import UnionDoc
 
-__version__ = "1.19.2"
+__version__ = "1.20.0b0"
 __all__ = [
     # ODM
     "Document",