+++ beanie/odm/settings/document.py
@@ -1,71 +1,10 @@
 from typing import Optional, List
 
 from pydantic import Field
-from pymongo import IndexModel
 
+from beanie.odm.fields import IndexModelField
 from beanie.odm.settings.base import ItemSettings
 from beanie.odm.settings.timeseries import TimeSeriesConfig
-
-
-class IndexModelField:
-    def __init__(self, index: IndexModel):
-        self.index = index
-        self.name = index.document["name"]
-
-        self.fields = tuple(sorted(self.index.document["key"]))
-        self.options = tuple(
-            sorted(
-                (k, v)
-                for k, v in self.index.document.items()
-                if k not in ["key", "v"]
-            )
-        )
-
-    def __eq__(self, other):
-        return self.fields == other.fields and self.options == other.options
-
-    def __repr__(self):
-        return f"IndexModelField({self.name}, {self.fields}, {self.options})"
-
-    @classmethod
-    def __get_validators__(cls):
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, v):
-        if isinstance(v, IndexModel):
-            return IndexModelField(v)
-        else:
-            return IndexModelField(IndexModel(v))
-
-    @staticmethod
-    def list_difference(
-        left: List["IndexModelField"], right: List["IndexModelField"]
-    ):
-        result = []
-        for index in left:
-            if index not in right:
-                result.append(index)
-        return result
-
-    @staticmethod
-    def list_to_index_model(left: List["IndexModelField"]):
-        return [index.index for index in left]
-
-    @classmethod
-    def from_motor_index_information(cls, index_info: dict):
-        result = []
-        for name, details in index_info.items():
-            fields = details["key"]
-            if ("_id", 1) in fields:
-                continue
-
-            options = {k: v for k, v in details.items() if k != "key"}
-            index_model = IndexModelField(
-                IndexModel(fields, name=name, **options)
-            )
-            result.append(index_model)
-        return result
 
 
 class DocumentSettings(ItemSettings):
@@ -77,6 +16,7 @@ class DocumentSettings(ItemSettings):
     single_root_inheritance: bool = False
 
     indexes: List[IndexModelField] = Field(default_factory=list)
+    merge_indexes: bool = False
     timeseries: Optional[TimeSeriesConfig] = None
 
     lazy_parsing: bool = False