+++ beanie/odm/utils/projection.py
@@ -11,10 +11,10 @@ def get_projection(
     model: Type[ProjectionModelType],
 ) -> Optional[Dict[str, int]]:
     if hasattr(model, "get_model_type") and (
-        model.get_model_type() == ModelType.UnionDoc
+        model.get_model_type() == ModelType.UnionDoc  # type: ignore
         or (  # type: ignore
-            model.get_model_type() == ModelType.Document
-            and model._inheritance_inited
+            model.get_model_type() == ModelType.Document  # type: ignore
+            and model._inheritance_inited  # type: ignore
         )
     ):  # type: ignore
         return None